+++ Tests/Standard.AspNetCore.TestApplication/Standard.AspNetCore.TestApplication.Infrastructure/HttpClients/IntegrationServiceProxyHttpClient.cs
@@ -120,7 +120,7 @@ namespace Standard.AspNetCore.TestApplication.Infrastructure.HttpClients
             httpRequest.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
 
             var content = JsonSerializer.Serialize(param1, _serializerOptions);
-            httpRequest.Content = new StringContent(content, Encoding.Default, "application/json");
+            httpRequest.Content = new StringContent(content, Encoding.UTF8, "application/json");
 
             using (var response = await _httpClient.SendAsync(httpRequest, HttpCompletionOption.ResponseHeadersRead, cancellationToken).ConfigureAwait(false))
             {
@@ -225,7 +225,11 @@ namespace Standard.AspNetCore.TestApplication.Infrastructure.HttpClients
                 using (var contentStream = await response.Content.ReadAsStreamAsync(cancellationToken).ConfigureAwait(false))
                 {
                     var str = await new StreamReader(contentStream).ReadToEndAsync(cancellationToken).ConfigureAwait(false);
-                    if (str != null && (str.StartsWith(@"""") || str.StartsWith("'"))) { str = str.Substring(1, str.Length - 2); };
+
+                    if (str.StartsWith(@"""") || str.StartsWith("'"))
+                    {
+                        str = str.Substring(1, str.Length - 2);
+                    }
                     return Guid.Parse(str);
                 }
             }
@@ -247,7 +251,11 @@ namespace Standard.AspNetCore.TestApplication.Infrastructure.HttpClients
                 using (var contentStream = await response.Content.ReadAsStreamAsync(cancellationToken).ConfigureAwait(false))
                 {
                     var str = await new StreamReader(contentStream).ReadToEndAsync(cancellationToken).ConfigureAwait(false);
-                    if (str != null && (str.StartsWith(@"""") || str.StartsWith("'"))) { str = str.Substring(1, str.Length - 2); }
+
+                    if (str.StartsWith(@"""") || str.StartsWith("'"))
+                    {
+                        str = str.Substring(1, str.Length - 2);
+                    }
                     return str;
                 }
             }
@@ -269,7 +277,11 @@ namespace Standard.AspNetCore.TestApplication.Infrastructure.HttpClients
                 using (var contentStream = await response.Content.ReadAsStreamAsync(cancellationToken).ConfigureAwait(false))
                 {
                     var str = await new StreamReader(contentStream).ReadToEndAsync(cancellationToken).ConfigureAwait(false);
-                    if (str != null && (str.StartsWith(@"""") || str.StartsWith("'"))) { str = str.Substring(1, str.Length - 2); };
+
+                    if (str.StartsWith(@"""") || str.StartsWith("'"))
+                    {
+                        str = str.Substring(1, str.Length - 2);
+                    }
                     return int.Parse(str);
                 }
             }