+++ Tests/Standard.AspNetCore.TestApplication/Standard.AspNetCore.TestApplication.Infrastructure/HttpClients/InvoiceServiceProxyHttpClient.cs
@@ -39,7 +39,7 @@ namespace Standard.AspNetCore.TestApplication.Infrastructure.HttpClients
             httpRequest.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
 
             var content = JsonSerializer.Serialize(dto, _serializerOptions);
-            httpRequest.Content = new StringContent(content, Encoding.Default, "application/json");
+            httpRequest.Content = new StringContent(content, Encoding.UTF8, "application/json");
 
             using (var response = await _httpClient.SendAsync(httpRequest, HttpCompletionOption.ResponseHeadersRead, cancellationToken).ConfigureAwait(false))
             {
@@ -51,7 +51,11 @@ namespace Standard.AspNetCore.TestApplication.Infrastructure.HttpClients
                 using (var contentStream = await response.Content.ReadAsStreamAsync(cancellationToken).ConfigureAwait(false))
                 {
                     var str = await new StreamReader(contentStream).ReadToEndAsync(cancellationToken).ConfigureAwait(false);
-                    if (str != null && (str.StartsWith(@"""") || str.StartsWith("'"))) { str = str.Substring(1, str.Length - 2); };
+
+                    if (str.StartsWith(@"""") || str.StartsWith("'"))
+                    {
+                        str = str.Substring(1, str.Length - 2);
+                    }
                     return Guid.Parse(str);
                 }
             }
@@ -104,7 +108,7 @@ namespace Standard.AspNetCore.TestApplication.Infrastructure.HttpClients
             httpRequest.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
 
             var content = JsonSerializer.Serialize(dto, _serializerOptions);
-            httpRequest.Content = new StringContent(content, Encoding.Default, "application/json");
+            httpRequest.Content = new StringContent(content, Encoding.UTF8, "application/json");
 
             using (var response = await _httpClient.SendAsync(httpRequest, HttpCompletionOption.ResponseHeadersRead, cancellationToken).ConfigureAwait(false))
             {